pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ga4gh-mcp-server"
version = "0.1.0"
description = "MCP server exposing the GA4GH ecosystem to AI clients"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.21.1",
    "ruff>=0.4.0",
    "coverage>=7.0.0",
]

[project.scripts]
ga4gh-mcp = "ga4gh_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ga4gh_mcp"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/ga4gh_mcp"]
